algorithm/leetcode/daily/1046-last-stone-weight_test.go
package daily

import "testing"

func Test_lastStoneWeight(t *testing.T) {
	type args struct {
		stones []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: "1", args: args{stones: []int{2, 7, 4, 1, 8, 1}}, want: 1},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := lastStoneWeight(tt.args.stones); got != tt.want {
				t.Errorf("lastStoneWeight() = %v, want %v", got, tt.want)
			}
		})
	}
}

func BenchmarkLastStoneWeight(b *testing.B) {
	n := 10000
	arr := make([]int, 0, n)
	for i := 0; i < n; i++ {
		arr = append(arr, i)
	}
	for i := 0; i < b.N; i++ {
		lastStoneWeight2(arr)
	}
}
